feat: Allow providing `customise` closure in `leptos_fluent!`
Currently, the only way to provide a `customise` closure for a fluent
bundle is by directly using `fluent_templates::static_loader!` and
providing the resulting translation bundle to `leptos_fluent!` in the
`translations` field. However, the `translations` field is optional
now and may be removed in the future depending on how
https://github.com/mondeja/leptos-fluent/issues/361 is implemented.

This PR adds support for providing the `customise` closure to
`leptos_fluent!` instead of needing to provide it via
`fluent_templates::static_loader!`.

Closes https://github.com/mondeja/leptos-fluent/issues/364